pyproject.toml
[project]
name = "morphogens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "equinox>=0.13.0",
    "jaxlib==0.5.1",
    "jupyter>=1.1.1",
    "jupyterlab>=4.4.4",
    "matplotlib>=3.10.3",
    "optax>=0.2.5",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
